+++ src/main/java/com/programming/techie/springngblog/common/UserConstant.java
@@ -2,8 +2,8 @@ package com.programming.techie.springngblog.common;
 
 public class UserConstant {
 
+    /****** DEFINITION OF THE ROLES ******/
     public static final String DEFAULT_ROLE = "USER";
     public static final String ADMIN_ACCESS = "ADMIN";
-    //public static final String[] MODERATOR_ACCESS = {"ROLE_MODERATOR"};
 
 }
