+++ src/main/java/com/programming/techie/springngblog/config/SecurityConfig.java
@@ -53,8 +53,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers("/api/events/**")
                 .permitAll()
                 .antMatchers("/api/admin/**")
-                .permitAll()
-                //.hasAuthority("ADMIN")
+                .hasAuthority("ADMIN")
                 .antMatchers("/post-photos/**")
                 .permitAll()
                 .anyRequest()
