+++ src/main/java/com/programming/techie/springngblog/dto/LoginRequest.java
@@ -1,22 +1,10 @@
 package com.programming.techie.springngblog.dto;
 
+import lombok.Data;
+
+@Data
 public class LoginRequest {
     private String username;
     private String password;
 
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
 }
