+++ src/main/java/com/programming/techie/springngblog/model/Categorie.java
@@ -1,7 +1,6 @@
 package com.programming.techie.springngblog.model;
 
 import lombok.Data;
-
 import javax.persistence.*;
 
 @Entity
@@ -12,9 +11,6 @@ public class Categorie {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
     private String name;
-
-
 
 }
