+++ src/main/java/com/programming/techie/springngblog/model/Post.java
@@ -4,8 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
-
 import javax.persistence.*;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotEmpty;
 import java.time.LocalDateTime;
 
 @Entity
@@ -14,23 +15,26 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class Post {
 
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-   // @NotBlank
+
+    @NotBlank
     @Column
     private String title;
+
     @Column
-   // @NotBlank
     private String username;
+
     @Lob
     @Column
-   // @NotEmpty
+    @NotEmpty
     private String content;
+
     @Column(name = "created_on", updatable = false, nullable = false)
     @CreationTimestamp
     private LocalDateTime createdOn;
+
     @Column(name = "updated_on")
     @UpdateTimestamp
     private LocalDateTime updatedOn;
@@ -38,7 +42,6 @@ public class Post {
     @OneToOne(fetch=FetchType.EAGER)
     private Categorie categorie;
 
-   // @Column(length= 45, nullable = true)
     private String image;
 
     public Post(){
@@ -51,7 +54,6 @@ public class Post {
         if (this.createdOn == null ){
             this.createdOn = LocalDateTime.now();
         }
-
     }
 
 }
