+++ src/main/java/com/programming/techie/springngblog/repository/CategorieRepository.java
@@ -4,6 +4,5 @@ import com.programming.techie.springngblog.model.Categorie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategorieRepository  extends JpaRepository<Categorie, Long> {
-
 }
