+++ src/main/java/com/programming/techie/springngblog/repository/PostRepository.java
@@ -7,6 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
-
     Optional<Post> findByUsername(String username);
 }
