+++ src/main/java/com/programming/techie/springngblog/security/MessageResponse.java
@@ -1,18 +1,13 @@
 package com.programming.techie.springngblog.security;
 
+import lombok.Data;
+
+@Data
 public class MessageResponse {
 
     private String message;
 
     public MessageResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
         this.message = message;
     }
 }
