+++ src/main/java/com/programming/techie/springngblog/service/CategorieService.java
@@ -1,8 +1,6 @@
 package com.programming.techie.springngblog.service;
 
-import com.programming.techie.springngblog.dto.PostDto;
 import com.programming.techie.springngblog.model.Categorie;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
