+++ src/main/java/com/programming/techie/springngblog/service/impl/PostServiceImpl.java
@@ -12,11 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.sql.Date;
-import java.time.Instant;
 import java.util.List;
-import java.util.Optional;
 
 import static java.util.stream.Collectors.toList;
 @Service
@@ -27,8 +23,7 @@ public class PostServiceImpl implements PostService {
     private AuthService authService;
     @Autowired
     private PostRepository postRepository;
-    @Autowired
-    private UserRepository userRepository;
+
     private ModelMapper modelMapper = new ModelMapper();
 
     @Override
@@ -39,7 +34,6 @@ public class PostServiceImpl implements PostService {
 
         @Override
     public Post createPost(PostDto postDto) {
-
       //  Post post = mapFromDtoToPost(postDto);
         Post post = modelMapper.map(postDto, Post.class);
             User loggedInUser = authService.getCurrentUser().orElseThrow(() -> new IllegalArgumentException("User Not Found"));
@@ -88,43 +82,3 @@ public class PostServiceImpl implements PostService {
         return post;
     }*/
 }
-
-
-
- /*   @Override
-    public List<Post> findPostsByUsername(String username) {
-        Optional<com.programming.techie.springngblog.model.User> optionalUser = userRepository.findByUsername(username);
-        return optionalUser.get().getPostList();
-    }*/
-
-
-    /*    @Override
-    public com.programming.techie.springngblog.model.User savePost(Post post, String username){
-        Optional<com.programming.techie.springngblog.model.User> userOptional = userRepository.findByUsername(username);
-        com.programming.techie.springngblog.model.User user = null;
-        if(userOptional.isPresent()){
-            user = userOptional.get();
-            post.setUser(user);
-            user.getPosts().add(post);
-            return userRepository.save(user);
-        }
-        return user;
-    }*/
-
-  /*  public Post getPostsByUsername(String username){
-        return postRepository.findByUsername(username).orElseThrow(IllegalArgumentException::new);
-    }*/
-
-/*    @Override
-    public List<Post> findPostsByUsername(String username) {
-        Optional<com.programming.techie.springngblog.model.User> optionalUser = userRepository.findByUsername(username);
-
-        User user =null;
-
-            user.getPostList().add(book);
-            userRepo.save(user);
-        }
-        return user;
-
-        return optionalUser.get().getPostList();
-    }*/
